use off_t consistently

+++ src/pkgfile.h
@@ -44,8 +44,8 @@
 struct archive_read_buffer {
   char *line;
   char *line_offset;
-  size_t line_size;
-  size_t real_line_size;
+  off_t line_size;
+  off_t real_line_size;
 
   char *block;
   char *block_offset;
